# Optimize `where` when range's begin/end are equal