For PostgreSQL:
```
create table users(
	id serial primary key,
	position  integer not null
);

insert into users (position) values (generate_series(1, 1000000));
analyze;
```

**Without BETWEEN**
```
explain analyze SELECT * FROM users WHERE position = 100000

Seq Scan on users  (cost=0.00..16925.00 rows=1 width=8) (actual time=8.640..86.703 rows=1 loops=1)
  Filter: ("position" = 100000)
  Rows Removed by Filter: 999999
Planning time: 0.040 ms
Execution time: 86.718 ms
```

**With BETWEEN**
```
explain analyze SELECT * FROM users WHERE position between 100000 and 100000;

Seq Scan on users  (cost=0.00..19425.00 rows=1 width=8) (actual time=8.511..105.254 rows=1 loops=1)
  Filter: (("position" >= 100000) AND ("position" <= 100000))
  Rows Removed by Filter: 999999
Planning time: 0.051 ms
Execution time: 105.268 ms
```

Approximately the same time differences are present for other values too.